requirements.txt
fastapi
uvicorn[standard]
deepgram-sdk
twilio
python-multipart
